# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Mock Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/token" : {
      "post" : {
        "tags" : [ "Login" ],
        "operationId" : "GetToken",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/search/autocomplete" : {
      "get" : {
        "tags" : [ "Search" ],
        "operationId" : "AutoComplete",
        "parameters" : [ {
          "name" : "term",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/tags" : {
      "get" : {
        "tags" : [ "Tags" ],
        "operationId" : "GetTags",
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      },
      "post" : {
        "tags" : [ "Tags" ],
        "operationId" : "CreateTag",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewTag"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewTag"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewTag"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/tags/{id}" : {
      "put" : {
        "tags" : [ "Tags" ],
        "operationId" : "UpdateTag",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DocTag"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DocTag"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DocTag"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      },
      "delete" : {
        "tags" : [ "Tags" ],
        "operationId" : "DeleteTag",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "NewTag" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "nullable" : true
          },
          "color" : {
            "type" : "string",
            "nullable" : true
          },
          "match" : {
            "type" : "string",
            "nullable" : true
          },
          "matching_algorithm" : {
            "type" : "integer",
            "format" : "int64"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "is_inbox_tag" : {
            "type" : "boolean"
          }
        },
        "additionalProperties" : false
      },
      "DocTag" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "slug" : {
            "type" : "string",
            "nullable" : true
          },
          "name" : {
            "type" : "string",
            "nullable" : true
          },
          "color" : {
            "type" : "string",
            "nullable" : true
          },
          "match" : {
            "type" : "string",
            "nullable" : true
          },
          "matching_algorithm" : {
            "type" : "integer",
            "format" : "int64"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "is_inbox_tag" : {
            "type" : "boolean"
          },
          "document_count" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "additionalProperties" : false
      },
      "UserInfo" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string",
            "nullable" : true
          },
          "password" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : false
      },
...
    }
  }
}